+++ src/Git_Updater/Additions/Repo_List_Table.php
@@ -80,6 +80,7 @@ class Repo_List_Table extends WP_List_Table {
 			$option['primary_branch']  = ! empty( $option['primary_branch'] ) ? $option['primary_branch'] : 'master';
 			$option['release_asset']   = ! empty( $option['release_asset'] ) ? '<span class="dashicons dashicons-yes"></span>' : false;
 			$option['private_package'] = ! empty( $option['private_package'] ) ? '<span class="dashicons dashicons-yes"></span>' : false;
+			$option['uses_lite']       = ! empty( $option['uses_lite'] ) ? '<span class="dashicons dashicons-yes"></span>' : false;
 
 			$options[ $key ] = $option;
 		}
@@ -119,12 +120,15 @@ class Repo_List_Table extends WP_List_Table {
 	public function column_default( $item, $column_name ) {
 		switch ( $column_name ) {
 			case 'uri':
+				return esc_url( (string) $item[ $column_name ] );
 			case 'slug':
 			case 'primary_branch':
-			case 'release_asset':
 			case 'type':
-			case 'private_package':
 				return esc_html( (string) $item[ $column_name ] );
+			case 'release_asset':
+			case 'private_package':
+			case 'uses_lite':
+				return (string) $item[ $column_name ];
 			default:
 				return esc_html( (string) wp_json_encode( $item ) );
 		}
@@ -215,6 +219,7 @@ class Repo_List_Table extends WP_List_Table {
 			'primary_branch'  => esc_html__( 'Primary Branch', 'git-updater' ),
 			'release_asset'   => esc_html__( 'Release Asset', 'git-updater' ),
 			'private_package' => esc_html__( 'Private Package', 'git-updater' ),
+			'uses_lite'       => esc_html__( 'Lite Client', 'git-updater' ),
 			'type'            => esc_html__( 'Type', 'git-updater' ),
 		];